requirements.txt
cycler==0.10.0
decorator==4.4.0
dimod==0.8.7
dwave-neal==0.4.5
dwave-networkx==0.7.1
GDAL==2.4.0
jsonschema==2.6.0
kiwisolver==1.0.1
lxml==4.3.2
matplotlib==3.0.3
networkx==2.2
nose==1.3.7
numpy==1.15.4
pandas==0.22.0
pkg-resources==0.0.0
pydot==1.4.1
pygraphviz==1.5
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
scipy==1.2.1
six==1.12.0
